pyproject.toml
[project]
name = "ai-codefix-assignment-nithesh"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.121.3",
  "python-dotenv>=1.2.1",
  "vllm==0.11.0",
]
